Removed local channels, added new textures for gates

+++ src/main/java/latmod/silicio/item/modules/ICBModule.java
@@ -10,7 +10,7 @@ public interface ICBModule // ItemModule
 	public void updateInvNet(CircuitBoard cb, int MID, FastList<InvEntry> list);
 	public void updateTankNet(CircuitBoard cb, int MID, FastList<TankEntry> list);
 	public CBChannel getChannel(CircuitBoard cb, int MID, int c);
-	public void setChannel(CircuitBoard cb, int MID, int c, byte ch);
+	public void setChannel(CircuitBoard cb, int MID, int c, int ch);
 	public int getChannelCount();
 	public IOType getModuleType();
 	public IOType getChannelType(int c);
